Checking in changes to allow hardware decoders to grab information from the header.
Now ORVDataDecoder has the functions:

    virtual std::string GetDictionaryObjectPath() {return "";}
    virtual void SetDecoderDictionary(const ORDecoderDictionary* aDict) {fDecoderDictionary=aDict;}
and the protected member:
    const ORDecoderDictionary* fDecoderDictionary;

ORDataProcessor dishes the DecoderDictionary to the decoder in SetDecoderDictionary() if the decoder overloads SetDecoderDictionary, 
otherwise fDecoderDictionary is set to NULL.

ORDecoderDictionary has the function

    virtual const ORDictionary* GetRecordDictWithCrateAndCard(int crate, int card) const;

to grab parameters of a record referencing a particular card/crate.

Oh and everything is backwards compatible.  
It will throw a warning if a decoder requests a dictionary and it can't find it, but otherwise no message will be given.


git-svn-id: svn://orca.physics.unc.edu/OrcaRoot@100 fd86f90f-2c0a-422a-b972-84106d8db31b